llm_scraper/requirements.txt
selenium>=4.10.0
webdriver-manager>=3.8.6
pandas>=1.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.1
tqdm>=4.65.0
beautifulsoup4>=4.12.0
playwright>=1.35.0
